requirements.txt
asyncio==3.4.3
aiofiles==23.1.0
apscheduler==3.6.3
aiohttp==3.8.4
alphabet-detector==0.0.7
bs4==0.0.1
bleach==6.0.0
beautifulsoup4==4.12.2
#profanity (By Gabi)
better-profanity==0.7.0
bing_image_downloader==1.1.2
cloudscraper==1.2.71
countryinfo
cachetools==4.2.2
CurrencyConverter
dateparser==1.1.8
dnspython
emoji==2.6.0
envparse==0.2.0
emoji-country-flag
future==0.18.3
fontTools
feedparser==6.0.10
ffmpeg-python==0.2.0
gtts
geopy==2.3.0
googletrans==4.0.0-rc1
google-trans-new==1.1.9
gpytranslate
hachoir==3.2.0
html2text==2020.1.16
jikanpy==4.3.2
lxml==4.9.2
motor==2.3.1
markdown2>=2.4.0
nudepy
nltk
nekos.py
opencv-python-headless==4.7.0.72
psycopg2-binary==2.9.6
pynewtonmath==1.0.1
pykeyboard==0.1.5
pendulum
python-telegram-bot==13.15
psutil==5.9.5
Pillow==9.5.0
pyrate-limiter==2.8.1
pyrogram==2.0.106
pytz
pySmartDL
python-arq==6.0.3
pretty_errors
# scanner requirements
PhoenixScanner==1.0.7
pymongo==3.12.3
regex==2023.6.3
redis
requests==2.31.0
spamwatch
speedtest-cli
sqlalchemy==1.3.20
spongemock
search_engine_parser
secureme
telegraph==2.1.0
telethon==1.24.0
textblob
TgCrypto
tswift
uvloop
ujson
wget
wikipedia==1.4.0
zalgo-text
unidecode
shortuuid
asyncio==3.4.3
pretty_errors==1.2.25
search_engine_parser==0.6.8
secureme==0.0.4
bing_image_downloader==1.1.2
